Test Project

+++ Arbolus.Model/Concrete/Expert.cs
@@ -6,7 +6,7 @@ namespace Arbolus.Model.Concrete
     {
         public string currency { get; set; }
 
-        public int hourlyRate { get; set; }
+        public decimal hourlyRate { get; set; }
 
         public string Name { get; set; }
